requirements.txt
crypto==1.4.1
cryptography==44.0.0
hpack==4.0.0
JsonForm==0.0.2
JsonSir==0.0.2
llvmlite==0.43.0
multidict==6.7.0
munkres==1.1.4
Naked==0.1.32
prettytable==3.10.0
pycryptodome==3.21.0
pynvml==11.5.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
Resource==0.2.1
sf_heu==0.5.2b0
shellescape==3.8.1
tenseal==0.3.15
toolz==1.1.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
ujson==5.10.0
